refactor(expo-linearlite): migrate to new multi-store API
