+++ examples/expo-linearlite/src/app/edit-issue.tsx
@@ -1,5 +1,5 @@
 import { queryDb } from '@livestore/livestore'
-import { useQuery, useStore } from '@livestore/react'
+import { useQuery } from '@livestore/react'
 import { Stack, useLocalSearchParams, useRouter } from 'expo-router'
 import { Image, Pressable, ScrollView, StyleSheet, TextInput, View } from 'react-native'
 
@@ -7,11 +7,12 @@ import { IssueStatusIcon, PriorityIcon } from '../components/IssueItem.tsx'
 import { ThemedText } from '../components/ThemedText.tsx'
 import { useThemeColor } from '../hooks/useThemeColor.ts'
 import { events, tables } from '../livestore/schema.ts'
+import { useAppStore } from '../livestore/store.ts'
 import type { Priority, Status } from '../types.ts'
 
 const EditIssueScreen = () => {
   const issueId = useLocalSearchParams().issueId as string
-  const { store } = useStore()
+  const store = useAppStore()
   const router = useRouter()
   const textColor = useThemeColor({}, 'text')
